refactor: cacheable_dyn macro support crate args
