+++ crates/rspack_cacheable_macros/src/cacheable_dyn/impl_disable.rs
@@ -1,0 +1,6 @@
+use proc_macro::TokenStream;
+
+/// impl cacheable dyn when disable
+pub fn impl_disable(input: TokenStream) -> TokenStream {
+  input
+}
